Move atlas SimpleServices to subpackage
